+++ hslua-core/test/HsLua/CoreTests.hs
@@ -34,6 +34,7 @@ import qualified Prelude
 import qualified Data.ByteString as B
 import qualified HsLua.Core.AuxiliaryTests
 import qualified HsLua.Core.ErrorTests
+import qualified HsLua.Core.PackageTests
 import qualified HsLua.Core.RunTests
 import qualified HsLua.Core.UnsafeTests
 import qualified HsLua.Core.UserdataTests
@@ -443,6 +444,7 @@ tests = testGroup "Core module"
       res <- run @Lua.Exception luaOp
       assertEqual "error handling leaks values to the stack" 0 res
 
+  , HsLua.Core.PackageTests.tests
   , HsLua.Core.RunTests.tests
   , HsLua.Core.UnsafeTests.tests
   , HsLua.Core.UserdataTests.tests