+++ hslua-packaging/test/HsLua/Packaging/ModuleTests.hs
@@ -16,7 +16,7 @@ import HsLua.Marshalling
   (Result (Success), peekIntegral, peekString, pushIntegral, pushText, runPeek)
 import HsLua.Packaging.Function
 import HsLua.Packaging.Module
-import Test.Tasty.HsLua ((=:), pushLuaExpr, shouldBeResultOf)
+import Test.Tasty.HsLua ((=:), shouldBeResultOf)
 import Test.Tasty (TestTree, testGroup)
 
 import qualified HsLua.Core as Lua
@@ -24,43 +24,7 @@ import qualified HsLua.Core as Lua
 -- | Specifications for Attributes parsing functions.
 tests :: TestTree
 tests = testGroup "Module"
-  [ testGroup "requirehs"
-    [ "pushes module to stack" =:
-      1 `shouldBeResultOf` do
-        Lua.openlibs
-        old <- Lua.gettop
-        Lua.requirehs "foo" (Lua.pushnumber 5.0)
-        new <- Lua.gettop
-        return (new - old)
-
-    , "module can be loaded with `require`" =:
-      let testModule = "string as a module"
-      in Just testModule `shouldBeResultOf` do
-        Lua.openlibs
-        Lua.requirehs "test.module" (Lua.pushstring testModule)
-        pushLuaExpr "require 'test.module'"
-        Lua.tostring Lua.top
-    ]
-
-  , testGroup "preloadhs"
-    [ "does not modify the stack" =:
-      0 `shouldBeResultOf` do
-        Lua.openlibs
-        old <- Lua.gettop
-        Lua.preloadhs "foo" (1 <$ Lua.pushnumber 5.0)
-        new <- Lua.gettop
-        return (new - old)
-
-    , "module can be loaded with `require`" =:
-      let testModule = "string as a module"
-      in Just testModule `shouldBeResultOf` do
-        Lua.openlibs
-        Lua.preloadhs "test.module" (1 <$ Lua.pushstring testModule)
-        pushLuaExpr "require 'test.module'"
-        Lua.tostring Lua.top
-    ]
-
-  , testGroup "creation helpers"
+  [ testGroup "creation helpers"
     [ "create produces a table" =:
       Lua.TypeTable `shouldBeResultOf` do
         Lua.newtable